+++ src/main/java/project/travelmate/response/PlanCreateResponse.java
@@ -1,6 +1,8 @@
 package project.travelmate.response;
 
 import lombok.Getter;
+import project.travelmate.domain.Plan;
+import project.travelmate.domain.PlanImage;
 
 @Getter
 public class PlanCreateResponse {
@@ -8,9 +10,9 @@ public class PlanCreateResponse {
     private Long planId;
     private String thumbnailUrl;
 
-    public PlanCreateResponse(Long planId, String thumbnailUrl) {
-        this.planId = planId;
-        this.thumbnailUrl = thumbnailUrl;
+    public PlanCreateResponse(Plan plan, PlanImage planImage) {
+        this.planId = plan.getId();
+        this.thumbnailUrl = planImage.getFilePath();
     }
 
 }
